+++ src/main/java/net/felsstudio/fels/lib/StringValue.java
@@ -32,7 +32,7 @@ public final class StringValue implements Value {
                 final Value[] chars = new Value[length()];
                 int i = 0;
                 for (char ch : value.toCharArray()) {
-                    chars[i++] = NumberValue.of((int) ch);
+                    chars[i++] = NumberValue.of(ch);
                 }
                 return new ArrayValue(chars);
             }
